Add simulation visualiser code
Adds a new file, visualiseSim that takes a .h5
file and visualises the data that it contains.
The permutation .txt file should also be supplied
with the same name.
Further changes include updates to the code that
embeds patterns in larger meshes, and writes
the meshfiles used by Chaste. Need to check the
change made to createFibroPattern